Move requests and responses to seperate files.
