+++ goosebit/api/devices/models.py
@@ -1,74 +1,13 @@
 from __future__ import annotations
 
-import asyncio
 import time
 from enum import StrEnum
 from typing import Annotated
 
-from fastapi.requests import Request
-from pydantic import BaseModel, BeforeValidator, Field, computed_field
+from pydantic import BaseModel, BeforeValidator, computed_field
 
 from goosebit.models import Device, UpdateModeEnum, UpdateStateEnum
 from goosebit.updater.manager import get_update_manager
-
-
-class DeviceTableResponse(BaseModel):
-    data: list[DeviceModel]
-    draw: int
-    records_total: int = Field(serialization_alias="recordsTotal")
-    records_filtered: int = Field(serialization_alias="recordsFiltered")
-
-    @classmethod
-    async def parse(cls, request: Request, query, search_filter, total_records):
-        params = request.query_params
-
-        draw = int(params.get("draw", 1))
-        start = int(params.get("start", 0))
-        length = int(params.get("length", 10))
-        search_value = params.get("search[value]", None)
-        order_column_index = params.get("order[0][column]", None)
-        order_column = params.get(f"columns[{order_column_index}][data]", None)
-        order_dir = params.get("order[0][dir]", None)
-
-        if search_value:
-            query = query.filter(search_filter(search_value))
-
-        if order_column:
-            query = query.order_by(f"{"-" if order_dir == "desc" else ""}{order_column}")
-
-        filtered_records = await query.count()
-        devices = await query.offset(start).limit(length).all()
-        data = list(await asyncio.gather(*[DeviceModel.parse(d) for d in devices]))
-
-        return cls(data=data, draw=draw, records_total=total_records, records_filtered=filtered_records)
-
-
-class DeviceAllResponse(BaseModel):
-    devices: list[DeviceModel]
-
-    @classmethod
-    async def parse(cls, devices: list[Device]):
-        return cls(devices=await asyncio.gather(*[DeviceModel.parse(d) for d in devices]))
-
-
-class UpdateDevicesRequest(BaseModel):
-    devices: list[str]
-    firmware: str | None = None
-    name: str | None = None
-    pinned: bool | None = None
-    feed: str | None = None
-
-
-class LogsDeviceResponse(BaseModel):
-    log: str | None
-
-
-class DeleteDevicesRequest(BaseModel):
-    devices: list[str]
-
-
-class ForceUpdateDevicesRequest(BaseModel):
-    devices: list[str]
 
 
 class UpdateStateModel(StrEnum):
